Remove background pruning from _deduplicate_message
This code was not Python 3 compatible due to the use of sys.maxint,
which was removed. This change simplifies the cleanup logic to only
when the function is called. This change makes test coverage more
complete.  It also removes a race conditon, that in principle could
cause messages that are not duplicate to be flagged as such. (This
will occur when execution stops just before `self._expiration_timer
= None` and a _deduplicate_message is executed, then no message is
received for EXCHANGE_LIFETIME, before a message is received.)